Move type definitions to single file
