--- src/blockchain.json
@@ -1,116 +1,0 @@
-[
-  {
-    "name": "Block",
-    "struct": [
-      { "name": "schemaVersion", "type": "Uint16" },
-      { "name": "proposerId", "type": "Uint16" },
-      { "name": "height", "type": "Uint64" },
-      { "name": "txCount", "type": "Uint32" },
-      { "name": "prevHash", "type": "Hash" },
-      { "name": "txHash", "type": "Hash" },
-      { "name": "stateHash", "type": "Hash" }
-    ]
-  },
-  {
-    "name": "TimeSpec",
-    "struct": [
-      { "name": "secs", "type": "Uint64" },
-      { "name": "nanos", "type": "Uint32" }
-    ]
-  },
-  {
-    "name": "Precommit",
-    "message": {
-      "serviceId": 0,
-      "messageId": 4,
-      "body": {
-        "struct": [
-          { "name": "validator", "type": "Uint16" },
-          { "name": "height", "type": "Uint64" },
-          { "name": "round", "type": "Uint32" },
-          { "name": "proposeHash", "type": "Hash" },
-          { "name": "blockHash", "type": "Hash" },
-          { "name": "time", "type": "TimeSpec" }
-        ]
-      }
-    }
-  },
-
-  {
-    "name": "ListProofNode",
-    "factory": {
-      "typeParams": [{ "name": "T" }],
-      "union": [
-        {
-          "name": "branch",
-          "type": {
-            "struct": [
-              { "name": "left", "type": { "ListProofNode": { "typeParam": "T" } } },
-              { "name": "right", "type": { "ListProofNode": { "typeParam": "T" } } }
-            ]
-          }
-        },
-        { "name": "stub", "type": { "ListProofNode": { "typeParam": "T" } } },
-        { "name": "hash", "type": "Hash" },
-        { "name": "val", "type": { "typeParam": "T" } }
-      ]
-    }
-  },
-
-  {
-    "name": "Bits256",
-    "struct": [
-      { "name": "isTerminal", "type": "Bool" },
-      { "name": "bytes", "type": "Hash" },
-      { "name": "bitLengthByte", "type": "Uint8" }
-    ]
-  },
-  {
-    "name": "MapProofNode",
-    "factory": {
-      "typeParams": [{ "name": "V" }],
-      "union": [
-        {
-          "name": "branch",
-          "type": {
-            "struct": [
-              { "name": "left", "type": { "MapProofNode": { "typeParam": "V" } } },
-              { "name": "right", "type": { "MapProofNode": { "typeParam": "V" } } },
-              { "name": "leftKey", "type": "Bits256" },
-              { "name": "rightKey", "type": "Bits256" }
-            ]
-          }
-        },
-        { "name": "hash", "type": "Hash" },
-        { "name": "val", "type": { "typeParam": "V" } }
-      ]
-    }
-  },
-  {
-    "name": "MapProofRoot",
-    "factory": {
-      "typeParams": [{ "name": "V" }],
-      "union": [
-        { "name": "empty", "type": "None" },
-        {
-          "name": "stub",
-          "type": {
-            "struct": [
-              { "name": "key", "type": "Bits256" },
-              {
-                "name": "value",
-                "type": {
-                  "union": [
-                    { "name": "hash", "type": "Hash" },
-                    { "name": "val", "type": { "typeParam": "V" } }
-                  ]
-                }
-              }
-            ]
-          }
-        },
-        { "name": "tree", "type": { "MapProofNode": { "typeParam": "V" } } }
-      ]
-    }
-  }
-]
