+++ src/index.js
@@ -1,10 +1,10 @@
 import * as crypto from './crypto'
-import types from './blockchain'
+import std from './std'
 
 // Export standard + blockchain types and resolver
 // XXX: Is there any way to do this in ES6 manner?
-for (let name in types) {
-  exports[name] = types[name]
+for (let name in std) {
+  exports[name] = std[name]
 }
 
 export { isExonumFactory, isExonumType, isExonumObject } from './lowlevel/common'
